fixed one bug, hit another
I carefully logged the segment a label is declared in but forgot to
actually save it in the table. This has been a theoretic concern for
some time, but I've never seen it actually happen until now. SubX is
just too low level.

Now I get past the first two phases but code generation fails to find
the 'Entry' label.