5665
The core library is a mess right now, because it can't use syntax sugar.
I need some way to tier the library so that later layers can use syntax
sugar, but without adding cognitive load.

A shell pipeline is pretty, but maybe I need to drop it. Just add functions
to layers. Get all of the translator out of the apps/ sub-directory.