5746
After much struggle, one more test: emitting a primitive with a register
operand.

The following two instructions have the same name:
  reg <- increment
  increment var

and they emit the same opcodes:
  ff 0/subop

But they're considered distinct policies in the code-generation 'table',
one for incrementing variables on the stack and the other for incrementing
variables in a register.