5748
A new draft of docs as I build out mu.subx. Once we have the second Mu
language, it makes sense to relegate most of the underlying SubX docs to
a separate doc.

I want each layer to be learned bottom-up as I've been organizing SubX
so far. But it's counter-productive to require people to learn SubX before
Mu.