5812 - debugging an ancient niggle in browse_trace
The problem: when I hit 'G' to go to the bottom of the trace, if the bottom
is visible on screen, the screen scrolls so the bottom of the trace
is the bottom-most line on screen. But the cursor moves to where the trace
used to end rather than the new location of the bottom of the trace (the
bottom of the screen).