6009 - significantly cleaner lexing
This cleans up a bunch of little warts that had historically accumulated
because of my bull-headedness in not designing a grammar up front. Let's
see if the lack of a grammar comes up again.

We now require that there be no space in variable declarations between
the name and the colon separating it from its type.