6017 - simplify type-dispatch for primitives
We'll be doing type-checking in a separate phase in future. For now we
need only to distinguish between literals and non-literals for x86 primitive
instructions.

I was tempted to support x86 set__ instructions for this change:
  https://c9x.me/x86/html/file_module_x86_id_288.html

That will happen at some point. And I'll simplify a bunch of branches for
results of predicate functions when it happens.