start migrating handles to fat pointers
CI will fail from this commit onward. Currently working:
  $ bootstrap translate init.linux 0[4-7]*.subx 080zero-out.subx -o a.elf  &&  ./a.elf test
  $ bootstrap run a.elf test
  $ chmod +x a.elf;  ./a.elf test

Plan: migrate functions that used to return handles to pass in a new arg
of type (addr handle). That's a bit of a weird type. There should be few
of these functions. (Open question: do we even want to expose this type
in the Mu language?)

Functions that just need to read from heap without modifying the handle
will receive `(addr T)` or `(handle T)` types as arguments.

As I sanitize each new file, I need to update signatures for any new functions
and add them to a list. I also need to update calls to any functions on
the list.