mu.subx: 25 failing tests remaining
This was again hard to debug, because:
  - the mapping between recursive calls to emit-subx-stmt-list and the
    nested blocks of code in the output kept blowing my stack, and
  - I kept expecting the problem to be there, or in the data attached to
    variables. But it was in the `emit-cleanup-code-until-depth` for the
    previous/inner block, in code I thought I'd already fixed (`clean-up-blocks`)
    that accidentally worked for previous tests.

Part of the problem here is that my library for the stack sucks. It's only
for 4-byte elements, but I'm using it for 8-byte elements (handles). Which
leads to double-pushes and double-pops and -- here -- an insidiously wrong
call to `top`.

Anyways: halfway there!