6449 - italics
We're not going to render italics since they still feel like an advanced
feature for terminals, and different terminals have different escape sequences
for them, and since they often look weird to my eyes on the monospace font
of a terminal window. So underscores and italics will both be bold.