6455 - abortive second attempt at parsing headings
This 'OO' approach seems more scalable, but I need to first wrestle with
a parsing issue: this text has a heading:

  abc *def
  # ghi*

Ugh, so I can't do this translation in a single pass.

Then again, maybe I should just keep going and not try to be compatible
with GitHub-Flavored Markdown. Require that new headings are also new paragraphs.