6670 - generic functions
Function signatures can now take type parameters starting with '_'.

Type parameters in a signature match any concrete type in the call. But
they have to be consistent within a single call.

Things I considered but punted on for now:
- having '_' match anything without needing to be consistent. Wildcards
  actually seem harder to understand.
- disallowing top-level '_' types. I'll wait until a concrete use case
  for disallowing.

We still don't support *defining* types with type parameters, so for now
this is only useful for calling functions on arrays or streams or handles.