6827 - tile: make top of stack line up
Idea from the Future of Software group. I was resistant for several days,
because stacks should grow up, and because it didn't seem like a problem
that the first column's result grew farther and farther away from its input.
But eventually I noticed that the final result can also be arbitrarily
far below, which seems more serious.

This approach also has the salubrious effect of not requiring the precomputation
of max-depth.