7389 - snapshot
Ok, we're back at the reset loop. Let's keep going; maybe having a decent
keyboard handler will fix it.

The bug I fixed here was caused by misunderstanding what m16&32 mean in
the Intel manual. It's still a regular regmem operand that uses all of
the ModR/M byte (which can be interpreted in 16-bit mode, adding to the
complication). It's just constrained to not allow direct addressing (mod 00).

I needed to better internalize the format of the instruction set references
at the start of Volume 2, Chapter 3.