7435 - baremetal: buffer for keyboard events
I'm trying to read the status register, but I'm still not seeing the breakpoint
being hit a second time. (And I again ran into the Bochs bug that breakpoints
at the first instruction of an interrupt handler don't work.)

Maybe this is just a debugger issue. Let's keep going, and try to start
using the keyboard events.